requirements.txt
# Core app dependencies
fastapi
uvicorn[standard]

# Whisper ASR
openai-whisper

# Sentence Transformers
sentence-transformers

# TTS - OpenVoice (ensure you have the openvoice package or local clone)
openvoice @ git+https://github.com/myshell-ai/OpenVoice.git

# VAD and audio handling
webrtcvad
soundfile

# Audio processing
torchaudio
numpy

# Model inference
transformers
torch

# Optional: for improved performance (GPU systems only)
accelerate


# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
